Make the GPUI stderr snapshots independent of backtrace presence
The Linux lane failed on `failing_scenario_diagnostic_is_written_to_stderr`.
Its snapshot captures a child process's stderr, and the panic hook prints
either a backtrace or the "run with `RUST_BACKTRACE=1`" note depending on the
ambient environment. The instrumented `--all-features --all-targets` build
retains debuginfo and the coverage run exports a backtrace setting, so frames
appeared where the previous lane had none.

Fix it at the seam rather than by loosening the assertion. The child process
now runs with `RUST_BACKTRACE=0`, and the snapshot settings strip both the
backtrace block and the note, so the captured stderr is identical whichever
form the hook chooses. The diagnostic itself, the feature path, the scenario
name, and the feature line all stay asserted verbatim.

Normalizing to the note rather than away from it does not work: the standard
library prints that note only once per process, so the second panic of the
teardown double-panic test has no note to match. Stripping both forms is the
only normalization that holds for one panic and for two.

Verified by running the whole snapshot binary under `RUST_BACKTRACE` unset, 0,
1, and full, and by removing each mechanism in turn to confirm the filters and
the child environment are independently sufficient.

The test module crossed the 400-line limit, so the snapshot redaction policy
and the tracing recorder move to `tests/support/mod.rs`, matching the
convention already used by the other crates in the workspace.

Also pin `RUST_BACKTRACE=0` on the three coverage steps so every lane captures
the same panic-hook output.
