+++ .github/workflows/ci.yml
@@ -445,6 +445,11 @@ jobs:
         uses: leynos/shared-actions/.github/actions/generate-coverage@7d46a399558914f5a05074e55a560fec0269fd0d
         env:
           RUN_RUST_CARGO_WAIT_TIMEOUT: '1800'
+          # Tests that snapshot a child process's stderr must see the same
+          # panic-hook output on every lane. An instrumented build retains
+          # debuginfo, so a backtrace would otherwise appear here and not on
+          # the Windows lanes.
+          RUST_BACKTRACE: '0'
         with:
           cache-provider: external
           output-path: coverage.xml
@@ -472,6 +477,8 @@ jobs:
           # Windows coverage runs (build plus doctests) exceed the shared
           # action's 600s per-invocation default; give them headroom.
           RUN_RUST_CARGO_WAIT_TIMEOUT: '1800'
+          # Match the Linux lane so stderr snapshots stay deterministic.
+          RUST_BACKTRACE: '0'
         with:
           cache-provider: external
           output-path: coverage.xml
@@ -490,6 +497,8 @@ jobs:
           # Windows coverage runs (build plus doctests) exceed the shared
           # action's 600s per-invocation default; give them headroom.
           RUN_RUST_CARGO_WAIT_TIMEOUT: '1800'
+          # Match the Linux lane so stderr snapshots stay deterministic.
+          RUST_BACKTRACE: '0'
         with:
           cache-provider: external
           output-path: coverage.xml
